Address request from #10771
There was a request to add a timeout to the join in #10771.
The thread itself is unlikely to hang - more likely situation here
is the subscription failing to cause updates, causing the main
thread to hang forever on the condition variable. Added time outs
to both so we can have a bit more information about what actually
happened when the test completes.